+++ src/com/booking/Models/alojamiento/Apartamento.java
@@ -12,7 +12,7 @@ public class Apartamento extends Alojamiento {
 
     }
 
-    public Apartamento(String nombre, String ciudad, TipoAlojamiento tipo, Float precioBase, List<Habitacion> habitaciones, Float calificacion, Date fechaInicio, Date fechaFin) {
+    public Apartamento(String nombre, String ciudad, TipoAlojamiento tipo, Double precioBase, List<Habitacion> habitaciones, Integer calificacion, Date fechaInicio, Date fechaFin) {
         super(nombre, ciudad, tipo, precioBase, habitaciones, calificacion, fechaInicio, fechaFin);
     }
