+++ src/com/booking/Models/alojamiento/Hotel.java
@@ -12,7 +12,7 @@ public class Hotel extends Alojamiento {
 
     }
 
-    public Hotel(String nombre, String ciudad, TipoAlojamiento tipo, Float precioBase, List<Habitacion> habitaciones, Float calificacion, Date fechaInicio, Date fechaFin) {
+    public Hotel(String nombre, String ciudad, TipoAlojamiento tipo, Double precioBase, List<Habitacion> habitaciones, Integer calificacion, Date fechaInicio, Date fechaFin) {
         super(nombre, ciudad, tipo, precioBase, habitaciones, calificacion, fechaInicio, fechaFin);
     }
